add non-verified versions of managed handlers
